Simple refactor: suggetor inside brain
