+++ internal/brain/suggestor_test.go
@@ -1,8 +1,7 @@
-package suggestor
+package brain
 
 import "testing"
 
 func TestSuggestMove(t *testing.T) {
 	t.Skip("Test me sometime")
 }
-
